--- a/library/src/androidTest/java/org/xmtp/android/library/ConversationsTest.kt
+++ b/library/src/androidTest/java/org/xmtp/android/library/ConversationsTest.kt
@@ -1,5 +1,6 @@
 package org.xmtp.android.library
 
+import android.util.Log
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -21,6 +22,7 @@ import org.xmtp.android.library.messages.createDeterministic
 import org.xmtp.android.library.messages.getPublicKeyBundle
 import org.xmtp.android.library.messages.toPublicKeyBundle
 import org.xmtp.android.library.messages.walletAddress
+import org.xmtp.proto.keystore.api.v1.Keystore
 import java.lang.Thread.sleep
 import java.util.Date
 
@@ -137,5 +139,47 @@ class ConversationsTest {
         }
 
         assertEquals(allMessages.size, 15)
+    }
+
+    @Test
+    fun testBigWallet() {
+        val privateKeyData = listOf(0x08, 0x36, 0x20, 0x0f, 0xfa, 0xfa, 0x17, 0xa3, 0xcb, 0x8b, 0x54, 0xf2, 0x2d, 0x6a, 0xfa, 0x60, 0xb1, 0x3d, 0xa4, 0x87, 0x26, 0x54, 0x32, 0x41, 0xad, 0xc5, 0xc2, 0x50, 0xdb, 0xb0, 0xe0, 0xcd)
+            .map { it.toByte() }
+            .toByteArray()
+        // Use hardcoded privateKey
+        val privateKey = PrivateKeyBuilder.buildFromPrivateKeyData(privateKeyData)
+        val privateKeyBuilder = PrivateKeyBuilder(privateKey)
+        val options =
+            ClientOptions(api = ClientOptions.Api(env = XMTPEnvironment.DEV))
+        val client = Client().create(account = privateKeyBuilder, options = options)
+
+        runBlocking {
+            val start = Date()
+            val conversations = client.conversations.list()
+            val end = Date()
+            Log.d("LOPI", "Loaded ${conversations.size} conversations in ${(end.time - start.time) / 1000.0}s")
+
+
+            val start2 = Date()
+            val conversations2 = client.conversations.list()
+            val end2 = Date()
+            Log.d("LOPI", "Second time loaded ${conversations2.size} conversations in ${(end2.time - start2.time) / 1000.0}s")
+
+            val last500Topics = conversations.takeLast(2000).map { it.toTopicData().toByteString() }
+            val client2 = Client().create(account = privateKeyBuilder, options = options)
+            for (topic in last500Topics) {
+                client2.conversations.importTopicData(Keystore.TopicMap.TopicData.parseFrom(topic))
+            }
+
+            val start3 = Date()
+            val conversations3 = client2.conversations.list()
+            val end3 = Date()
+            Log.d("LOPI", "Loaded ${conversations3.size} conversations in ${(end3.time - start3.time) / 1000.0}s")
+
+            val start4 = Date()
+            val conversations4 = client2.conversations.list()
+            val end4 = Date()
+            Log.d("LOPI", "Second time loaded ${conversations4.size} conversations in ${(end4.time - start4.time) / 1000.0}s")
+        }
     }
 }