Outputs
```
Loaded 2015 conversations in 17.609s
Second time loaded 2015 conversations in 0.197s
Loaded 2015 conversations in 0.326s
Second time loaded 2015 conversations in 0.192s
```