@nplasterer 2024-04-03
Turned out to just be an iOS bug.